+++ packages/StateChange/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+## Added
+- String option for property name; [#15](https://github.com/domxjs/domx/issues/15)
+
+## Updated
+- Logging to include the state property name
+
+## Breaking Changes
+- [#15](https://github.com/domxjs/domx/issues/15) Updated option names from
+`prop` to `property` and from `changeEventName` to `changeEvent`.
 
 ## [0.5.0] - 2021-08-16
 ### Added